Move some files to better locations
