+++ packages/backend/src/peripherals/database/WithdrawableBalanceChange.ts
@@ -1,7 +1,7 @@
 import { AssetHash, EthereumAddress, StarkKey } from '@explorer/types'
 
-import { assertUnreachable } from '../../../utils/assertUnreachable'
-import { ToJSON } from './ToJSON'
+import { assertUnreachable } from '../../utils/assertUnreachable'
+import { ToJSON } from './transactions/ToJSON'
 
 export type WithdrawableBalanceChangeData =
   | WithdrawalAllowedData
